requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
jinja2==3.1.3
python-multipart==0.0.6
resend==0.7.0
python-dotenv==1.0.0
